requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.3
django-cleanup==8.1.0
djangorestframework==3.15.1
fonttools==4.51.0
fpdf2==2.7.8
idna==3.7
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
